# src/repositories/classroom/dto/response-classroom.dto.ts
import { Type } from 'class-transformer';
import { IsBoolean, IsNotEmpty, IsOptional, IsString } from 'class-validator';

import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

import { Department } from '../../department/entities';
import { Classroom } from '../entities';

export class ResponseClassroomDto {
  @ApiProperty()
  @IsNotEmpty()
  @Type(() => Number)
  id: number;

  @ApiProperty()
  @IsNotEmpty()
  @IsString()
  name: string;

  @ApiPropertyOptional()
  @IsOptional()
  @IsString()
  description?: string;

  @ApiProperty()
  @IsNotEmpty()
  @IsString()
  type: string;

  @ApiProperty()
  @IsNotEmpty()
  @Type(() => Number)
  departmentIds: number[];

  @ApiProperty()
  @IsOptional()
  @Type(() => Department)
  departments: Department[];

  @ApiProperty()
  @IsNotEmpty()
  @IsBoolean()
  status!: boolean;

  constructor(data: Classroom) {
    this.id = data.id;
    this.name = data.name;
    this.description = data.description;
    this.status = data.status;
    this.departments = data.departments;
    this.departmentIds = data?.departments?.map((department) => department.id);
    this.type = data.type;
  }
}