# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/career": {
      "post": {
        "operationId": "CareerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCareerDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCareerDto"
                }
              }
            }
          }
        },
        "tags": [
          "career"
        ]
      },
      "get": {
        "operationId": "CareerController_findAll",
        "parameters": [
          {
            "name": "schoolId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseCareerDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "career"
        ]
      }
    },
    "/career/{id}": {
      "get": {
        "operationId": "CareerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCareerDto"
                }
              }
            }
          }
        },
        "tags": [
          "career"
        ]
      },
      "patch": {
        "operationId": "CareerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCareerDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCareerDto"
                }
              }
            }
          }
        },
        "tags": [
          "career"
        ]
      },
      "delete": {
        "operationId": "CareerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseCareerDto"
                }
              }
            }
          }
        },
        "tags": [
          "career"
        ]
      }
    },
...
    "/day": {
      "post": {
        "operationId": "DayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDayDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDayDto"
                }
              }
            }
          }
        },
        "tags": [
          "day"
        ]
      },
      "get": {
        "operationId": "DayController_findAll",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseDayDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "day"
        ]
      }
    },
    "/day/{id}": {
      "get": {
        "operationId": "DayController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDayDto"
                }
              }
            }
          }
        },
        "tags": [
          "day"
        ]
      },
      "patch": {
        "operationId": "DayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDayDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDayDto"
                }
              }
            }
          }
        },
        "tags": [
          "day"
        ]
      },
      "delete": {
        "operationId": "DayController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDayDto"
                }
              }
            }
          }
        },
        "tags": [
          "day"
        ]
      }
    },
...
    "/classroom": {
      "post": {
        "operationId": "ClassroomController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseClassroomDto"
                }
              }
            }
          }
        },
        "tags": [
          "classroom"
        ]
      },
      "get": {
        "operationId": "ClassroomController_findAll",
        "parameters": [
          {
            "name": "classroomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseClassroomDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "classroom"
        ]
      }
    },
    "/classroom/{id}": {
      "get": {
        "operationId": "ClassroomController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseClassroomDto"
                }
              }
            }
          }
        },
        "tags": [
          "classroom"
        ]
      },
      "patch": {
        "operationId": "ClassroomController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassroomDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseClassroomDto"
                }
              }
            }
          }
        },
        "tags": [
          "classroom"
        ]
      },
      "delete": {
        "operationId": "ClassroomController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseClassroomDto"
                }
              }
            }
          }
        },
        "tags": [
          "classroom"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/recovery-password": {
      "post": {
        "operationId": "AuthController_generateRecovery",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryPasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/create-student": {
      "post": {
        "operationId": "AuthController_createStudent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/recovery-password/{recovery_token}": {
      "get": {
        "operationId": "AuthController_getRecoveryById",
        "parameters": [
          {
            "name": "recovery_token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "AuthController_postRecoveryById",
        "parameters": [
          {
            "name": "recovery_token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "put": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SIGEIT-API Documentation",
    "description": "API completa para el Sistema de Gestión de Información Educativa y Tecnológica",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateCareerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "abbreviation",
          "department",
          "status"
        ]
      },
      "Department": {
        "type": "object",
        "properties": {}
      },
      "ResponseCareerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "departmentId": {
            "type": "number"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "abbreviation",
          "departmentId",
          "department",
          "status"
        ]
      },
      "UpdateCareerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
      "School": {
        "type": "object",
        "properties": {}
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "idDocument": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "teacher": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "status": {
            "type": "boolean"
          },
          "school": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "department": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "career": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "email",
          "role",
          "idDocument",
          "name",
          "teacher",
          "status",
          "school",
          "department",
          "career"
        ]
      },
      "Teacher": {
        "type": "object",
        "properties": {}
      },
      "Career": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDayDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "abbreviation",
          "status"
        ]
      },
      "ResponseDayDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "abbreviation",
          "status"
        ]
      },
      "UpdateDayDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
      "CreateClassroomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "departments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdCreateEntity"
            }
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "type",
          "departments",
          "status"
        ]
      },
      "ResponseClassroomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "departmentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "departmentIds",
          "departments",
          "status"
        ]
      },
      "UpdateClassroomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "departments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IdCreateEntity"
            }
          },
          "status": {
            "type": "boolean"
          }
        }
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginUserResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "loginStamp": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "teacher": {
            "$ref": "#/components/schemas/Teacher"
          },
          "career": {
            "$ref": "#/components/schemas/Career"
          }
        },
        "required": [
          "email",
          "id",
          "name",
          "role",
          "loginStamp",
          "school",
          "department",
          "teacher",
          "career"
        ]
      },
      "RecoveryPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "RecoveryPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}